pyproject.toml
[project]
name = "tranquil-turing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]
